Use size_t for array length

+++ src/libmongoc/src/mongoc/mongoc-crypto-cng.c
@@ -159,11 +159,11 @@ _crypto_hash_size (mongoc_crypto_t *crypto)
 static int
 _bcrypt_derive_key_pbkdf2 (BCRYPT_ALG_HANDLE prf,
                            const char *password,
-                           uint32_t password_len,
+                           size_t password_len,
                            const uint8_t *salt,
-                           uint32_t salt_len,
+                           size_t salt_len,
                            uint32_t iterations,
-                           uint32_t key_len,
+                           size_t key_len,
                            unsigned char *output)
 {
    // Make non-const versions of password and salt.
@@ -192,11 +192,11 @@ _bcrypt_derive_key_pbkdf2 (BCRYPT_ALG_HANDLE prf,
 static void
 mongoc_crypto_cng_derive_key_pbkdf2 (mongoc_crypto_t *crypto,
                                      const char *password,
-                                     uint32_t password_len,
+                                     size_t password_len,
                                      const uint8_t *salt,
-                                     uint32_t salt_len,
+                                     size_t salt_len,
                                      uint32_t iterations,
-                                     uint32_t key_len,
+                                     size_t key_len,
                                      unsigned char *output)
 {
    uint8_t intermediate_digest[MONGOC_SCRAM_HASH_MAX_SIZE];
@@ -224,11 +224,11 @@ mongoc_crypto_cng_derive_key_pbkdf2 (mongoc_crypto_t *crypto,
 int
 mongoc_crypto_cng_pbkdf2_hmac_sha1 (mongoc_crypto_t *crypto,
                                     const char *password,
-                                    uint32_t password_len,
+                                    size_t password_len,
                                     const uint8_t *salt,
-                                    uint32_t salt_len,
+                                    size_t salt_len,
                                     uint32_t iterations,
-                                    uint32_t key_len,
+                                    size_t key_len,
                                     unsigned char *output)
 {
 #if defined(MONGOC_HAVE_BCRYPT_PBKDF2)
@@ -274,11 +274,11 @@ mongoc_crypto_cng_sha1 (mongoc_crypto_t *crypto,
 int
 mongoc_crypto_cng_pbkdf2_hmac_sha256 (mongoc_crypto_t *crypto,
                                       const char *password,
-                                      uint32_t password_len,
+                                      size_t password_len,
                                       const uint8_t *salt,
-                                      uint32_t salt_len,
+                                      size_t salt_len,
                                       uint32_t iterations,
-                                      uint32_t key_len,
+                                      size_t key_len,
                                       unsigned char *output)
 {
 #if defined(MONGOC_HAVE_BCRYPT_PBKDF2)
