+++ src/libmongoc/src/mongoc/mongoc-crypto-common-crypto-private.h
@@ -30,11 +30,11 @@ BSON_BEGIN_DECLS
 int
 mongoc_crypto_common_crypto_pbkdf2_hmac_sha1 (mongoc_crypto_t *crypto,
                                               const char *password,
-                                              uint32_t password_len,
+                                              size_t password_len,
                                               const uint8_t *salt,
-                                              uint32_t salt_len,
+                                              size_t salt_len,
                                               uint32_t iterations,
-                                              uint32_t key_len,
+                                              size_t key_len,
                                               unsigned char *output);
 
 void
@@ -54,11 +54,11 @@ mongoc_crypto_common_crypto_sha1 (mongoc_crypto_t *crypto,
 int
 mongoc_crypto_common_crypto_pbkdf2_hmac_sha256 (mongoc_crypto_t *crypto,
                                                 const char *password,
-                                                uint32_t password_len,
+                                                size_t password_len,
                                                 const uint8_t *salt,
-                                                uint32_t salt_len,
+                                                size_t salt_len,
                                                 uint32_t iterations,
-                                                uint32_t key_len,
+                                                size_t key_len,
                                                 unsigned char *output);
 
 void
